- Rename store/configureStore to store/default - Add missing import for redux.compose
