+++ app/client/main.js
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
 import App from './containers/App';
-import configureStore from './store/configureStore'
+import configureStore from './store/default';
 
 const initialState = {};
 const store = configureStore(initialState);
